requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic>=2.6.4,<3.0.0
transformers==4.34.0
torch==2.3.1
sentence-transformers==2.2.2
numpy==1.26.1
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
openai>=1.27.0
anthropic>=0.3.0
pytest==7.0.1
httpx>=0.21.0
open-interpreter==0.3.4
sqlmodel>=0.0.19